requirements.txt
aiohttp==3.5.4
aniso8601==6.0.0
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
chardet==3.0.4
Click==7.0
discord.py==1.0.1
Flask==1.0.2
Flask-Cors==3.0.7
flask-marshmallow==0.10.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.2
multidict==4.5.2
pep8==1.7.1
pkg-resources==0.0.0
pycodestyle==2.5.0
PyMySQL==0.9.3
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
typing-extensions==3.7.2
websockets==6.0
Werkzeug==0.15.2
yarl==1.3.0
